webApp/requirements.txt
flask
flask-login
flask-mysqldb
flask-WTF
mysql-connector-python